Dockerfile changes to handle onnxruntime dependency (#793)
* Update Dockerfile

* update onnx installation, and update tests

* update test onnx utils
